feat: add GoHomeButton component and "Go back" button on dashboard
Extract shared go-home logic (warning modal + stop processes + navigate)
into a reusable GoHomeButton component. Used in both the NavBar header
and the Dashboard title row with different icons (House vs ArrowLeft).
