+++ src/components/ui/index.ts
@@ -1,5 +1,6 @@
 export { ErrorFallback } from "./ErrorFallback";
 export { FadeIn } from "./FadeIn";
+export { GoHomeButton } from "./GoHomeButton";
 export { LoadingRing } from "./LoadingRing";
 export { Logo } from "./Logo";
 export { Modal } from "./Modal";
